Cleanup, inlining
